StereoToMono also in effects module, but should it remain an effect?
