+++ modules/effects/mod-effect-classes/CMakeLists.txt
@@ -73,6 +73,8 @@ list( APPEND SOURCES
    ScienFilter.h
    Silence.cpp
    Silence.h
+   StereoToMono.cpp
+   StereoToMono.h
    ToneGen.cpp
    ToneGen.h
    TruncSilence.cpp
