+++ modules/effects/mod-effect-classes/StereoToMono.h
